ui/nodes: Move NodePageMetricsTab from dumb component to container
Refs: #2776
